+++ tests/testthat/test-files.R
@@ -2,7 +2,6 @@ test_that("list works", {
   skip_if_offline()
   expect_true((n <- nrow(files <- sooty_files())) > 15000)
 
-  expect_true((nn <- nrow(files <- sooty_files(FALSE))) > n)
 
   expect_true(nrow(.curated_files("oisst-tif")) < n)
 })
@@ -24,3 +23,4 @@ test_that("deprecation message helpful", {
   skip_if_offline()
   expect_warning(dataset())
 })
+